Feat: Use AmritaSense
