+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "amrita_core"
-version = "0.8.6.1"
+version = "0.9.0rc1"
 description = "High performance, lightweight agent framework."
 readme = "README.md"
 requires-python = ">=3.10,<3.15"
@@ -8,11 +8,11 @@ dependencies = [
     "aiofiles>=25.1.0",
     "aiohttp>=3.13.3",
     "aiologic>=0.16.0",
-    "anthropic>=0.84.0",
+    "amrita-sense>=0.2.1",
     "anyio>=4.12.1",
     "fastmcp>=3.2.0", # Updated, see https://github.com/advisories/GHSA-rww4-4w9c-7733 for details.
     "filetype>=1.2.0",
-    "jieba>=0.42.1",
+    "importlib>=1.0.4",
     "jinja2>=3.1.6",
     "loguru>=0.7.3, <0.8.0",
     "openai>=2.16.0",
@@ -29,13 +29,19 @@ dependencies = [
 dev-dependencies = [
     "ruff>=0.12.8",
     "pytest>=8.4.1",
-    "mcp>=1.21.1",
     "py-spy>=0.4.1",
     "pyright>=1.1.407",
+    "anthropic>=0.84.0",
     "pytest-asyncio>=1.3.0",
+    "jieba>=0.42.1",
     "pytest-cov>=7.0.0",
 ]
 package = true
+
+[project.optional-dependencies]
+anthropic = ["anthropic>=0.84.0"]
+jieba = ["jieba>=0.42.1"]
+all = ["amrita_core[anthropic,jieba]"]
 
 [tool.ruff]
 line-length = 88
