+++ src/amrita_core/_env.py
@@ -1,0 +1,4 @@
+from amrita_sense.utils import Ref
+
+TEST_MODE: Ref[bool] = Ref(False)
+_MODULE_LOADED: dict[str, bool] = {}
