+++ src/amrita_core/hook/on.py
@@ -1,5 +1,6 @@
+from amrita_sense.hook.on import on_event
+
 from .event import EventTypeEnum
-from .matcher import Matcher
 
 
 def on_completion(priority: int = 10, block: bool = True):
@@ -12,7 +13,3 @@ def on_precompletion(priority: int = 10, block: bool = True):
 
 def on_preset_fallback(priority: int = 10, block: bool = True):
     return on_event(EventTypeEnum.PRESET_FALLBACK, priority, block)
-
-
-def on_event(event_type: EventTypeEnum | str, priority: int = 10, block: bool = True):
-    return Matcher(event_type, priority, block)
